// src/fin/types.d.ts
export type CW20Denom =
  | {
      native: string;
    }
  | {
      cw20: Addr;
    };
export type Addr = string;
export type Decimal = string;
export type Uint128 = string;
export interface BookResponse {
  base: PoolResponse[];
  quote: PoolResponse[];
}
export interface PoolResponse {
  offer_denom: CW20Denom;
  quote_price: Decimal;
  total_offer_amount: Uint128;
}
export type Precision =
  | { significant_figures: number }
  | { decimal_places: number };

export interface ConfigResponse {
  denoms: [CW20Denom, CW20Denom];
  is_bootstrapping: boolean;
  owner: Addr;
  price_precision: Precision;
  decimal_delta: number;
}
// ...
export type Timestamp = Uint64;
export type Uint64 = string;
export interface OrderResponse {
  created_at: Timestamp;
  filled_amount: Uint128;
  idx: Uint128;
  offer_amount: Uint128;
  offer_denom: CW20Denom;
  original_offer_amount: Uint128;
  owner: Addr;
  quote_price: Decimal;
}
export interface OrdersResponse {
  orders: OrderResponse[];
}
export interface PriceResponse {
  pools: [PoolResponse, PoolResponse];
}
// ...
export type AssetInfo =
  | {
      token: {
        contract_addr: string;
      };
    }
  | {
      native_token: {
        denom: string;
      };
    };
export interface Asset {
  amount: Uint128;
  info: AssetInfo;
}
export interface SimulationResponse {
  commission_amount: Uint128;
  return_amount: Uint128;
  spread_amount: Uint128;
}
export interface FinReadOnlyInterface {
  contractAddress: string;
  config: () => Promise<ConfigResponse>;
  simulation: ({
    offerAsset,
  }: {
    offerAsset: Asset;
  }) => Promise<SimulationResponse>;
  order: ({ orderIdx }: { orderIdx: string }) => Promise<OrderResponse>;
  ordersByUser: ({
    address,
    limit,
    startAfter,
  }: {
    address: string;
    limit?: number;
    startAfter?: Uint128;
  }) => Promise<OrdersResponse>;
  price: ({ price }: { price: string }) => Promise<PriceResponse>;
  book: ({
    limit,
    offset,
  }: {
    limit?: number;
    offset?: number;
  }) => Promise<BookResponse>;
}